Update access_checker.php
As it's theoretically impossible to reverse a string hashed with SHA1, I have to remove it for the RFID tag numbers in order to display them on the webpage. However, instead of removing SHA1 hashing for the RFID tags, I replaced it with base 64 encoding. Not as secured as SHA1 encryption, but it's still better than no encryption at all and in doesn't appear in clear text into the database.